Returnin False when it fails

+++ 2-do_deploy_web_static.py
@@ -6,7 +6,7 @@ from datetime import datetime
 from pathlib import Path
 
 env.hosts = ["34.229.70.28", "54.89.45.26"]
-env.exit_on_error = False
+#env.exit_on_error = False
 
 
 def do_pack():
@@ -27,28 +27,43 @@ def do_deploy(archive_path):
     try:
         with open(archive_path, "rb") as f:
             pass
-    except (FileNotFoundError, SystemExit):
+        archive_stem = Path(archive_path)
+        archive_stem = archive_stem.stem
+        put(archive_path, '/tmp/')
+        # Make Folder for where to copy to
+        result = sudo(f"mkdir -p /data/web_static/releases/{archive_stem}")
+        if result.failed:
+            return False
+        result = sudo(
+            f"tar -xzf /tmp/{archive_stem}.tgz\
+            -C /data/web_static/releases/{archive_stem}"
+            )
+        if result.failed:
+            return False
+        result = sudo(f"rm /tmp/{archive_stem}.tgz")
+        if result.failed:
+            return False
+        result = sudo(
+            f"mv -f /data/web_static/releases/{archive_stem}/web_static/*\
+            /data/web_static/releases/{archive_stem}/web_static/*/*\
+            /data/web_static/releases/{archive_stem}"
+            )
+        if result.failed:
+            return False
+        result = sudo(
+            f"rm -rf /data/web_static/releases/{archive_stem}/web_static"
+            )
+        if result.failed:
+            return False
+        result = sudo("rm -rf /data/web_static/current")
+        if result.failed:
+            return False
+        result = sudo(
+            f"ln -s /data/web_static/releases/{archive_stem}\
+            /data/web_static/current"
+            )
+        if result.failed:
+            return False
+        return (True)
+    except (FileNotFoundError):
         return False
-    archive_stem = Path(archive_path)
-    archive_stem = archive_stem.stem
-    put(archive_path, '/tmp/')
-    # Make Folder for where to copy to
-    run(f"mkdir -p /data/web_static/releases/{archive_stem}")
-    sudo(
-        f"tar -xzf /tmp/{archive_stem}.tgz\
-        -C /data/web_static/releases/{archive_stem}"
-        )
-    sudo(f"rm /tmp/{archive_stem}.tgz")
-    run(
-        f"mv -f /data/web_static/releases/{archive_stem}/web_static/*\
-        /data/web_static/releases/{archive_stem}"
-        )
-    run(
-        f"rm -rf /data/web_static/releases/{archive_stem}/web_static"
-        )
-    run("rm -rf /data/web_static/current")
-    run(
-        f"ln -s /data/web_static/releases/{archive_stem}\
-        /data/web_static/current"
-        )
-    return (True)
